+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Copy and pasting the git commit messages is __NOT__ enough.
 
 ## [Unreleased]
 ### Added
+- Added vehicle of interest coloring through scenario studio. This lets the scenario color vehicles that match a certain pattern of vehicle id.
+- SMARTS now provides `remove_provider` to remove a provider from the simulation. Use carefully.
 ### Changed
 ### Deprecated
 ### Fixed
@@ -21,7 +23,6 @@ Copy and pasting the git commit messages is __NOT__ enough.
 ### Added
 - Added support for the [Argoverse 2 Motion Forecasting Dataset](https://www.argoverse.org/av2.html#forecasting-link) (see `scenarios/argoverse`)
 - Added `waymo_open_dataset` as a module at the SMARTS repo level, to be able to load waymo scenarios without any external packages
-- Added vehicle of interest coloring through scenario studio. This lets the scenario color vehicles that match a certain pattern of vehicle id.
 ### Changed
 - Changed the `lanepoint_spacing` setting in `MapSpec` to be non-optional. Lanepoints are now generated lazily when waypoints are used.
 ### Deprecated