+++ CHANGELOG.md
@@ -10,11 +10,14 @@ Copy and pasting the git commit messages is __NOT__ enough.
 
 ## [Unreleased]
 ### Added
+- Added vehicle of interest coloring through scenario studio. This lets the scenario color vehicles that match a certain pattern of vehicle id.
+- SMARTS now provides `remove_provider` to remove a provider from the simulation. Use carefully.
 ### Changed
 ### Deprecated
 ### Fixed
 - Fixed "rl/racing" `numpy` incompatibility.
 - Fixed an issue with SUMO maps where waypoints in junctions would not return all possible paths.
+- Fixed an issue in Argoverse maps where adjacent lanes would sometimes not be grouped in the same road.
 ### Removed
 ### Security
 
@@ -22,7 +25,6 @@ Copy and pasting the git commit messages is __NOT__ enough.
 ### Added
 - Added support for the [Argoverse 2 Motion Forecasting Dataset](https://www.argoverse.org/av2.html#forecasting-link) (see `scenarios/argoverse`)
 - Added `waymo_open_dataset` as a module at the SMARTS repo level, to be able to load waymo scenarios without any external packages
-- Added vehicle of interest coloring through scenario studio. This lets the scenario color vehicles that match a certain pattern of vehicle id.
 ### Changed
 - Changed the `lanepoint_spacing` setting in `MapSpec` to be non-optional. Lanepoints are now generated lazily when waypoints are used.
 ### Deprecated