+++ CHANGELOG.md
@@ -10,6 +10,14 @@ Copy and pasting the git commit messages is __NOT__ enough.
 
 ## [Unreleased]
 ### Added
+### Changed
+### Deprecated
+### Fixed
+### Removed
+### Security
+
+## [1.0.10] # 2023-03-27
+### Added
 - Added vehicle of interest coloring through scenario studio. This lets the scenario color vehicles that match a certain pattern of vehicle id.
 - SMARTS now provides `remove_provider` to remove a provider from the simulation. Use carefully.
 ### Changed